create a folder for claim status tab and move files into that
