+++ src/applications/claims-status/components/ClaimsDecision.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Link } from 'react-router';
 import moment from 'moment';
 import PropTypes from 'prop-types';
+import NextSteps from './claim-status-tab/NextSteps';
 
 const formatDate = closedDate => moment(closedDate).format('MMMM D, YYYY');
 
@@ -10,7 +11,7 @@ const headerText = closedDate =>
 
 const ClaimsDecision = ({ completedDate, showClaimLettersLink }) => (
   <>
-    <va-alert uswds="false">
+    <va-alert uswds="false" class="vads-u-margin-y--2">
       <h3 className="claims-alert-header" slot="headline">
         {completedDate && headerText(completedDate)}
       </h3>
@@ -53,46 +54,7 @@ const ClaimsDecision = ({ completedDate, showClaimLettersLink }) => (
         </p>
       </va-alert>
     )}
-    <h4 className="claims-paragraph-header vads-u-font-size--h3">Next steps</h4>
-    <p>
-      <strong>If you agree with your claim decision</strong>, you don’t need to
-      do anything else.
-    </p>
-    <p>
-      <strong>If you have new evidence</strong> that shows your condition is
-      service connected, you can file a Supplemental Claim. We’ll review your
-      claim decision using the new evidence.
-      <br />
-      <a href="/decision-reviews/supplemental-claim/">
-        Learn more about Supplemental Claims
-      </a>
-    </p>
-    <p>
-      <strong>
-        If your condition has gotten worse since you filed your claim
-      </strong>
-      , you can file a new claim for an increase in disability compensation.
-      <br />
-      <a href="/disability/how-to-file-claim/">
-        Learn how to file a VA disability claim
-      </a>
-    </p>
-    <p>
-      <strong>If you disagree with your claim decision</strong>, you can request
-      a decision review.
-      <br />
-      <a href="/resources/choosing-a-decision-review-option/">
-        Find out how to choose a decision review option
-      </a>
-    </p>
-    <p>
-      <strong>If you’re not enrolled in VA health care</strong>, you can apply
-      now.
-      <br />
-      <a href="/health-care/apply/application/introduction">
-        Apply for VA health care benefits
-      </a>
-    </p>
+    <NextSteps />
   </>
 );
