+++ src/applications/claims-status/components/claim-status-tab/NextSteps.jsx
@@ -1,0 +1,52 @@
+import React from 'react';
+
+function NextSteps() {
+  return (
+    <div className="next-steps-container">
+      <h3 className="vads-u-margin-top--0 vads-u-margin-bottom--3">
+        Next Steps
+      </h3>
+      <p>
+        <strong>If you agree with your claim decision</strong>, you don’t need
+        to do anything else.
+      </p>
+      <p>
+        <strong>If you have new evidence</strong> that shows your condition is
+        service connected, you can file a Supplemental Claim. We’ll review your
+        claim decision using the new evidence.
+        <br />
+        <a href="/decision-reviews/supplemental-claim/">
+          Learn more about Supplemental Claims
+        </a>
+      </p>
+      <p>
+        <strong>
+          If your condition has gotten worse since you filed your claim
+        </strong>
+        , you can file a new claim for an increase in disability compensation.
+        <br />
+        <a href="/disability/how-to-file-claim/">
+          Learn how to file a VA disability claim
+        </a>
+      </p>
+      <p>
+        <strong>If you disagree with your claim decision</strong>, you can
+        request a decision review.
+        <br />
+        <a href="/resources/choosing-a-decision-review-option/">
+          Find out how to choose a decision review option
+        </a>
+      </p>
+      <p>
+        <strong>If you’re not enrolled in VA health care</strong>, you can apply
+        now.
+        <br />
+        <a href="/health-care/apply/application/introduction">
+          Apply for VA health care benefits
+        </a>
+      </p>
+    </div>
+  );
+}
+
+export default NextSteps;
